logger test placed to unit directory with trivial changes
